tests: add more tests
